--enh Adding fdescribe_implementation, xdescribe_implementation
+ hacking it to the `global`
