+++ jasmine.run.js
@@ -1,5 +1,6 @@
 import Jasmine from 'jasmine';
 import { SpecReporter } from 'jasmine-spec-reporter';
+import './tests/helpers/implementation_iterator';
 
 var jasmine = new Jasmine();
 jasmine.loadConfigFile('./jasmine.json');
