+++ tests/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "globals": {
+    "describe_implementation": "readonly"
+  }
+}
